Move Spinner to atoms
