+++ src/stories/atoms/icon.stories.svelte
@@ -1,5 +1,5 @@
 <script module>
-  import { default as Icon } from "../../ui/components/icon.svelte";
+  import { default as Icon } from "../../ui/atoms/icon.svelte";
   import { brandingModes } from "../../../.storybook/modes";
   import {
     type Args,
